Start coalescing state into a struct
The idea here is to create a "tree" of nested structures. This allows
us to code the application as a sequence of events that operate on
that tree.

If we notice that we are repeating parts of the tree, those parts can
become proper stand-alone structs with their own impls.

Just an experiment.