Harden collision detector
There were some edge cases with regards to numerical overflows
and underflows that would cause crashes. This patch fixes that
(mostly). It also makes the algorithm attempt to give reasonable
results when given unrealistic inputs.

I'm currently fuzzing the algorithm to see if there's anything else
that's wrong with it.