Implement the skeleton for the game shell
This commit adds a game_shell entry point and starts a game shell thread
when the game is started. The game shell takes in a logger (so it can
log messages) and will in the future take a channel to main and so
probably more.

The idea behind the GameShell is simple:

	The GameShell attaches to /run/universe-gsh.sock
	It uses Metac for each input line it gets and returns a result

Then, to actually send commands and receive feedback we have a new
program called `gsh` (GameShell) that attaches to that same socket.