gsh/main: Exit correctly when main tries to exit
The current gsh server `accept`s connections. This operation blocks.

What we do now since this patch is send a single TCP connection request
when we try to shut down. gsh then checks an atomic to see if it needs
to exit.

If the atomic is false, gsh exits gracefully.