logger: Implement more benchmarking
Currently the logger uses 9 ns/iter on a dynamically (globally) disabled
log level.

slog spends 6 ns/iter.

The reason for this is that slog uses a non-shared variable for checking
the log level, whereas logger uses an atomic variable. We _can_ do that
as well, but then the logger objects in other threads won't be able to use
that log level information.

Overall, I'm not too concerned, in the actual logging case we're 5x
faster anyway, which is where it mostly counts. Ease of use will go down
if we do not use an atomic, each logger would for instance periodically
check an atomic while having its own locally stored log level.

We could certainly try...
<time passes>
It works actually, but we now have a check for the atomic every 256th
logging invocation. So in real applications it can take a while (we lose
logs) for a logger to activate. So I'll forgo it here, it did reach 6
ns/iter just like with slog.