gsh: [Squashed] Move from array to hashmap based command handling
Partial commit - Get working SPEC parsing

Commit before it works

Get stuff to work (no cmd interpreting yet!)

We (sadly) have to re-initialize the spec after creating a thread. This
is because the Nest thing has a non-static lifetime. But it's fine, we
can live with this, and creating a connection is something that happens
so seldomly that it ought to be fine, connections are supposed to be
long-lived after all.

Get a semi-working lookup thing going

Still not quite working, having a hard time finding the actual fn after
a predicate... not sure what happens.

WOOT! Get something that works!