gsh: [Squashed] Improve command mapping and implement subcommands
gsh: Change the mapping from <X, Ether> to <str, (X, Ether)>

This is a VASTLY superior mapping, as it correctly makes X part of the
value, not the key, so we don't need to depend on rust nightly for the
function get_key_value(). After this I'll clean up some more.

Get recurseive parsing to work.

This patch makes () evaluate correctly, and sees the result as an atom.
This atom can be a string containing spaces. So if some function wants a
string with spaces, () can be used.