gsh: Correctly perform autocomplete
The current implementation has some rough edges, as it's a very crude
construction, but it works!

What's left to do: handling of Tcp streams as datagrams, we currently
rely on lucky timing to get a single command right (as we flush), but
this may change when we send a bunch of commands in rapid succession. On
both the client and server side we'll need some kind of line-based
parser. This _could_ be done inside metac by getting a `Reader` and
reading until the stream is empty.

It's possible, but it just doesn't sit right by me (metac taking control
of Reader is bad). Rather, we could have... an iterator... `.lines()`...
Doesn't that already exist?