logger: Make messages more readable, remove [] from context
I'd like to be as practical as possible when it comes to the log
messages. While we may write systems to parse them and see if something
is amiss, it's also useful for log messages to be easily parsable for
humans.

For instance, having MON DAY YEAR with the month in letters makes it
vastly easier to parse, in that order. Using the standard
2018-08-23T19:30:12.21304+01:00 is useful for machine communication, and
maybe we'll revert to that in the future, but right now this is quite
good.

The format right now is

<date> <time>: <lvl> <ctx>[ <optional-part>]: <content>

': ' is used to visually delimit where things are positioned, as without
it, the whole message seems too "fluid". It's hard to describe, but
people like to have boundaries, to get an indication of "here, this
thing stops and another begins".