metac: Add partial parsing
Partial parsing allows us to parse a stream of bytes with an unknown
future. This is especially useful for stream sockets like TCP, here, we
have no message boundary, so we must parse each byte that comes in and
determine if we've reached the end of a statement or not.

The other parsers (in interpret_single, interpret_multiple) assume
complete inputs (that is, complete statements.

interpret_multiple could be used on a file, and interpret_single on a
single statement, but they can not handle streams.

One thing that's different about PartialParser is that it's not part of
the Evaluate trait, and there's a reason:

	We do not know the nature of the stream, and can hence not
	"store" it appropriately.

So we leave storage to the caller, we just signal whether a statement
has been found, an error, or nothing interesting changed.
The caller is responsible for allocating the bytes on some buffer,
however they want.