draw: Get to draw two triangles!
The big idea is to now use a trait for the Surface type. I'll call it
something different from that to avoid a clash with hal, but let's call
it something like Canvas.

Now a Canvas trait should provide a framebuffer and a queue_group that
you can queue stuff to, in addition to a viewport (which we have to give
to drawing calls out of necessity).

This allows us to create arbitrary canvases, and holy cow it seems to
work with multiple triangles! That's awesome :D.

For the swapchain on the screen, we DO need special care, we need to
manually swap it, we should also add a dummy draw that draws nothing but
just signals a fence.

The fence can then be waited on in the drop. So during a drop, we should
always draw a final "thing" and then wait for a fence? Or should we wait
for the fence later?