draw: Separate device connection from screen
The device is internally synchronized in vulkan, so we can freely share
a reference to this in all objects, allowing each to deallocate their
memory. This also ensures that we can not deallocate the device before
having dropped all memory holders which is nice.

The API is a little more verbose right now but I hope to put everything
behind a nice simple struct.