draw: Bind descriptor set layouts when drawing textures
We need to bind descriptor set layouts so that the GPU uses the correct
texture when drawing, previously, the previous state leaked causing a
single textured thing to work, but with multiple textures we need to
take care of binding the set layout.

I might put the descriptor pool inside the device structure so we can
always get a new descriptor.