ketimer: Pass the time instant to the delegate/callback
The reason for doing this is that the delegate may be interested in the
time that has elapsed. In a concrete case, we need this in order for
some packets to be resent, not all packets will be resent, just those
that haven't received an acknowledgement within a certain amount of
time, hence, we send the time instant in.

No duration is sent in, as the delegate can remember this himself.