vxdraw: [WIP] Add debug triangles
Currently draws arbitrary triangles on screen. Instancing is NOT used
because of the simpleness of the mesh (instancing is slower).

For later, the debug triangles will have:

        Color
        Not be affected by transformations

These triangles can be useful to figuring out certain conditions, e.g. a
blue triangle indicates that we are slow or something. I don't know,
they should probably be renamed to debug triangles.

Finally, I want to make the API safer by providing a struct for the
triangle. It also uses `device.wait_idle` because we can't overwrite a
buffer when we're inside a draw call. It also needs a re-allocation
scheme.