vxdraw: Do not access the swapchan image after it has been presented
This was technically wrong, even though the validation layer didn't
notify of it. See https://www.khronos.org/registry/vulkan/specs/1.0-extensions/html/vkspec.html#_wsi_swapchain

Now we use a function call inserted between the submit and present. it
is up to the implementor to wait for the frame. This could alternatively
be signalled by using a data type, but that's for the future.