vxdraw: Make DebugTriangles a proper type
This patch removes the painful way we rotated and scaled the debug
triangles, before, we had to convert them back into origin triangles
before doing operations and translate them back. This is very lossy, as
floating point operations lose precision. So, instead there is now just
a static origin triangle in addition to a scale, a rotation, and a
translation, and that's fine. It's completely bindless.

We do spend about 2x the amount of data, but that should be OK for the
determinism we get.