vxdraw: Add model scaling to debug triangles
Model scaling scales each model so that it keeps its shape in a
non-square render window. This is much easier on the eyes. Depending on
the aspect ratio, we may scale x if the screen is wider or y if it is
taller. It is very important that the shape is only made smaller to
avoid potential collisions or overlap.

As it stands, the translation still occurs in the familiar -1..1 space,
but the model may be squeezed into a smaller model space.

This ensures that no matter what kind of window you have, if you add the
sum of the two objects' radii together when moving them, they will never
touch, well, they will _exactly_ touch (if you rotate them
appropriately).