vxdraw: Add a quad
Quads are useful for a variety of reasons. One can draw a screen-wide
quad to perform fading operations, or use a quad as a backdrop for a
texture, a border, or whatever. The reason for not using triangles is
because triangles are very cumbersome to handle, you'd need two
triangles. Quads abstract this out.