vxdraw/dyntex: Reuse removed sprite indices
We now reuse sprite indices that have been removed by pushing said index
to a vector, when we then push a sprite again, we check if there are any
available holes we can fill.

This should keep the drawing system alive a little longer (so we don't
overallocate).

Not drawing the holes is achieved by making them scale 0, kind-of a hack
but it's a compromise between access speed and draw speed. This favors
access speed, since we can't (well we can but it's a lot of work) change
the meaning of the indices handed out.