vxdraw: Start transitioning to methods
Using methods makes bundles of calls way shorter and more readable. The
current design is to for each type of drawable, we have a method to
access the entire class by proxy by using `.type()` on the `Windowing`
struct:

        let dyntex = windowing.dyntex();
        let id = dyntex.push_texture(...);
        dyntex.push_sprite(id, ...);
        dyntex.push_sprite(id, ...);
        dyntex.push_sprite(id, ...);
        dyntex.push_sprite(id, ...);
        dyntex.push_texture(...);

And so on, if we want to do operations on a strtex, we'll need to drop
the dyntex (because it holds a &mut Windowing), and then acquire a
strtex.

        let strtex = windowing.strtex();
        let id = strtex.push_texture(...);
        strtex.set_pixel(id, 300, 200, (255, 0, 255, 255));

We now avoid having the prefix in the name as:

        windowing.strtex_set_pixel

as well as avoiding direct child access as:

        windowing.strtex.set_pixel

This can still be done if needed:

        windowing.strtex().set_pixel

We'll move away from the free methods soon.