vxdraw/dyntex: Start work on dynamically resizable buffers
The way the vertex buffers are designed is to be fully filled during
each frame, thus, it makes the job of actually resizing quite easy. We
let `Vec` do the hard work of resizing to user input, and we only resize
to what's necessary when it's necessary. This should be fairly
efficient.

I also want one for indices now, since that will allow us to have
infinite bullets (dyntex sprites).