vxdraw/debtri: Use the mockbuffer specializations
This patch implements mockbuffers: all changes to debtris are now stores
in separate vectors per "thing", so there's a rotation vector, a
translation vector, and so on.

During each draw call, we copy the entire vector into the actual GPU
buffer, with a catch...

Every time we edit the mockbuffer, we reset a counter, every time we
draw, we decrement this counter. When the counter is zero, we _do not_
copy the mockbuffer to the GPU buffer (because it's not needed).

There's a significant speedup in benchmarks

Before:

        test debtri::tests::bench_rotating_windmills                                ... bench:     656,766 ns/iter (+/- 45,110)
        test debtri::tests::bench_rotating_windmills_no_render                      ... bench:      37,340 ns/iter (+/- 635)

After:

        test debtri::tests::bench_rotating_windmills                                ... bench:     428,388 ns/iter (+/- 23,084)
        test debtri::tests::bench_rotating_windmills_no_render                      ... bench:         219 ns/iter (+/- 4)

Most notably, editing itself has been sped up significantly because we
now access a _very_ tightly packed array in RAM instead of mapping each
change on the GPU buffer. The above is for 1000 items.

Similarly for 100_000: ~24ms vs 14.7 ms, so it seems to be good.