metac: Make interpret_multiple ignore whitespace lines
While we can explicitly interpret whitespace, for multiline commands we
want to ignore empty lines. The reason we want to ignore them is because
a function looking like:

        fn name (
                f
                g
        )

Would cause the last line with the `)` to be considered an empty-line
that would be interpreted. That's not good. This avoids that problem.

An alternative is for the evaluator to ignore empty commands, but then
it needs to cache the previous result, which is kinda wasteful.